unirev/requirement.txt
Django              2.1.1
bcrypt              3.1.4
Jinja2              2.10
mysqlclient         1.3.13
numpy               1.13.1
Pillow              5.2.0
PyMySQL             0.9.2
pytest              3.2.3
urllib3             1.22
cryptography        2.3.1
djangorestframework 3.9.0